Implement unsubscribe page
Not ideal for now, it would be better to handle
all of this server-side and then redirect the user
to a client page saying "completed with success"
